refactor: move tabular exceptions into tabular domain
